requirements.txt
asgiref==3.8.1
attrs==22.2.0
Django==4.2.11
djangorestframework==3.15.1
iniconfig==1.1.1
packaging==21.3
Pillow
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.1
sqlparse==0.5.0
toml==0.10.2
drf_spectacular
djoser
djangorestframework-simplejwt
drf-nested-routers==0.93.4
pyJWT==2.1.0